add seperate date functions from Table row component
