+++ frontend/src/components/ScheduledEventsTableRow/ScheduledEventsTableRow.tsx
@@ -1,4 +1,5 @@
 import React, { FunctionComponent } from 'react';
+import { deriveTimeFromStartDate, deriveDateFromStartDate } from './index';
 
 export interface ScheduledEventsTableRowProps {
   id: number;
@@ -12,29 +13,6 @@ export interface ScheduledEventsTableRowProps {
 }
 
 export const ScheduledEventsTableRow: FunctionComponent<ScheduledEventsTableRowProps> = (props) => {
-  const derivedTime = (): string => {
-    const dt = new Date();
-    dt.setTime(Date.parse(props.start_date));
-    let hours = dt.getUTCHours();
-    const AmOrPm = hours >= 12 ? 'PM' : 'AM';
-    hours = hours % 12 || 12;
-    const minutes = dt.getUTCMinutes();
-    const finalTime = hours + ':' + (minutes < 10 ? '0' + minutes : minutes) + ' ' + AmOrPm;
-
-    return finalTime;
-  };
-
-  const derivedDate = (): string => {
-    const day = new Date();
-    day.setTime(Date.parse(props.start_date));
-    const dd = String(day.getDate()).padStart(2, '0');
-    const mm = String(day.getMonth() + 1).padStart(2, '0');
-    const yyyy = day.getFullYear();
-    const newDate = dd + '-' + mm + '-' + yyyy;
-
-    return newDate;
-  };
-
   return (
     <tr>
       <td className="text-center">{props.id}</td>
@@ -49,7 +27,7 @@ export const ScheduledEventsTableRow: FunctionComponent<ScheduledEventsTableRowP
         </div>
       </td>
       <td>
-        {derivedDate()} at {derivedTime()}
+        {deriveDateFromStartDate(props)} at {deriveTimeFromStartDate(props)}
       </td>
       <td>{props.repeat ? 'Every ' + props.interval + ' ' + props.interval_type : 'None'}</td>
     </tr>
