+++ frontend/src/components/ScheduledEventsTableRow/index.ts
@@ -1,1 +1,2 @@
 export * from './ScheduledEventsTableRow';
+export * from './utils/dateConversion';
